enhance: simplify underlying WAL scanner ownership
